Add update item's image functionality.

+++ src/main/java/com/practicespring/printerstore/repositories/ItemRepository.java
@@ -3,5 +3,5 @@ package com.practicespring.printerstore.repositories;
 import com.practicespring.printerstore.models.Item;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ItemRepository extends CrudRepository<Item, Integer> {
+public interface ItemRepository extends CrudRepository<Item, Long> {
 }
